web: splitting the validate funcs to enable generation
