+++ azurerm/internal/services/web/validate/app_service_custom_hostname_binding_id.go
@@ -6,7 +6,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/web/parse"
 )
 
-func CustomHostnameBindingID(i interface{}, k string) (warnings []string, errors []error) {
+func AppServiceCustomHostnameBindingID(i interface{}, k string) (warnings []string, errors []error) {
 	v, ok := i.(string)
 	if !ok {
 		errors = append(errors, fmt.Errorf("expected type of %q to be string", k))
