+++ azurerm/internal/services/web/validate/app_service_environment_id.go
@@ -2,7 +2,6 @@ package validate
 
 import (
 	"fmt"
-	"regexp"
 
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/web/parse"
 )
@@ -17,16 +16,6 @@ func AppServiceEnvironmentID(i interface{}, k string) (warnings []string, errors
 	if _, err := parse.AppServiceEnvironmentID(v); err != nil {
 		errors = append(errors, fmt.Errorf("Can not parse %q as a resource id: %v", k, err))
 		return
-	}
-
-	return warnings, errors
-}
-
-func AppServiceEnvironmentName(v interface{}, k string) (warnings []string, errors []error) {
-	value := v.(string)
-
-	if matched := regexp.MustCompile(`^[0-9a-zA-Z][-0-9a-zA-Z]{0,61}[0-9a-zA-Z]$`).Match([]byte(value)); !matched {
-		errors = append(errors, fmt.Errorf("%q may only contain alphanumeric characters and dashes up to 60 characters in length, and must start and end in an alphanumeric", k))
 	}
 
 	return warnings, errors
