+++ azurerm/internal/services/web/validate/app_service_id.go
@@ -2,20 +2,9 @@ package validate
 
 import (
 	"fmt"
-	"regexp"
 
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/web/parse"
 )
-
-func AppServiceName(v interface{}, k string) (warnings []string, errors []error) {
-	value := v.(string)
-
-	if matched := regexp.MustCompile(`^[0-9a-zA-Z-]{1,60}$`).Match([]byte(value)); !matched {
-		errors = append(errors, fmt.Errorf("%q may only contain alphanumeric characters and dashes and up to 60 characters in length", k))
-	}
-
-	return warnings, errors
-}
 
 // AppServiceID validates that the specified ID is a valid App Service ID
 func AppServiceID(i interface{}, k string) (warnings []string, errors []error) {
