+++ azurerm/internal/services/web/validate/app_service_name.go
@@ -1,0 +1,16 @@
+package validate
+
+import (
+	"fmt"
+	"regexp"
+)
+
+func AppServiceName(v interface{}, k string) (warnings []string, errors []error) {
+	value := v.(string)
+
+	if matched := regexp.MustCompile(`^[0-9a-zA-Z-]{1,60}$`).Match([]byte(value)); !matched {
+		errors = append(errors, fmt.Errorf("%q may only contain alphanumeric characters and dashes and up to 60 characters in length", k))
+	}
+
+	return warnings, errors
+}
