+++ azurerm/internal/services/web/validate/app_service_plan_id.go
@@ -2,7 +2,6 @@ package validate
 
 import (
 	"fmt"
-	"regexp"
 
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/web/parse"
 )
@@ -18,16 +17,6 @@ func AppServicePlanID(i interface{}, k string) (warnings []string, errors []erro
 	if _, err := parse.AppServicePlanID(v); err != nil {
 		errors = append(errors, fmt.Errorf("Can not parse %q as a resource id: %v", k, err))
 		return
-	}
-
-	return warnings, errors
-}
-
-func AppServicePlanName(v interface{}, k string) (warnings []string, errors []error) {
-	value := v.(string)
-
-	if matched := regexp.MustCompile(`^[0-9a-zA-Z-_]{1,60}$`).Match([]byte(value)); !matched {
-		errors = append(errors, fmt.Errorf("%q may only contain alphanumeric characters, dashes and underscores up to 60 characters in length", k))
 	}
 
 	return warnings, errors
