+++ azurerm/internal/services/web/validate/app_service_plan_name.go
@@ -1,0 +1,16 @@
+package validate
+
+import (
+	"fmt"
+	"regexp"
+)
+
+func AppServicePlanName(v interface{}, k string) (warnings []string, errors []error) {
+	value := v.(string)
+
+	if matched := regexp.MustCompile(`^[0-9a-zA-Z-_]{1,60}$`).Match([]byte(value)); !matched {
+		errors = append(errors, fmt.Errorf("%q may only contain alphanumeric characters, dashes and underscores up to 60 characters in length", k))
+	}
+
+	return warnings, errors
+}
